pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "homelocal-auth"
version = "0.1.0"
description = "Reusable JWT authentication library for Home.Local FastAPI services"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Home.Local Platform Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "respx>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/homelocal/homelocal-auth"
Repository = "https://github.com/homelocal/homelocal-auth"

[tool.hatch.build.targets.wheel]
packages = ["homelocal_auth"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
